requirements.txt
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
idna==2.9
numpy==1.18.4
pandas==1.0.3
pathspec==0.8.0
psycopg2-binary==2.8.5
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
regex==2020.5.14
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.17
tableauhyperapi==0.0.10622
tableauserverclient==0.11
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
